Only author can edit and delete recipe  added.

+++ Kulinarna.Web/Pages/Delete.cshtml.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Kulinarna.BusinessLogic.DTOs;
 using Kulinarna.BusinessLogic.Services;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
@@ -12,11 +13,14 @@ namespace Kulinarna.Web.Pages
     public class DeleteModel : PageModel
     {
         private readonly RecipeService recipeService;
+        private readonly UserManager<ApplicationUser> userManager;
+
         public RecipeDTO RecipeDTO { get; set; }
 
-        public DeleteModel(RecipeService recipeService)
+        public DeleteModel(RecipeService recipeService, UserManager<ApplicationUser> userManager)
         {
             this.recipeService = recipeService;
+            this.userManager = userManager;
         }
         public IActionResult OnGet(int recipeId)
         {
@@ -31,7 +35,15 @@ namespace Kulinarna.Web.Pages
         {
             var recipe = recipeService.GetRecipeById(recipeId);
 
-            await recipeService.Delete(recipeId);
+            if (recipe.AuthorId != userManager.GetUserId(HttpContext.User)) 
+            {
+                return RedirectToPage("./404");
+            }
+
+            else
+            {
+                await recipeService.Delete(recipeId);
+            }
             if (recipe == null)
             {
                 return RedirectToPage("./404");
